added job executor

+++ models/jobmodels/job.go
@@ -12,7 +12,7 @@ type Job struct {
 	ID string `json:"id,omitempty" bson:"id,omitempty"`
 
 	// Trigger time in milliseconds
-	TriggerMS int             `json:"trigger_ms,omitempty" bson:"trigger_ms,omitempty"`
+	TriggerMS int64           `json:"trigger_ms,omitempty" bson:"trigger_ms,omitempty"`
 	Meta      json.RawMessage `json:"meta,omitempty" bson:"meta,omitempty"`
 	Route     string          `json:"route,omitempty" bson:"route,omitempty"`
 }
